third_party/WebKit/LayoutTests/fast/css/pseudo-not-empty-adjacent-dynamic-expected.txt

Check that the sibling of an element that becomes :not(:empty) is properly recalculated

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(sibling, '').backgroundColor is transparent
PASS getComputedStyle(sibling, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
